+++ application/classes/Core/ControllerName.php
@@ -2,14 +2,19 @@
 namespace Core {   
 	class ControllerName {
 		/**
-		* @var string The string to be used as method name
+		* @var string The string to be used as base
 		*/
 		private $string = '';
 
 		/**
-		* @var string Holds the sanitized method name.
+		* @var string Holds the sanitized name.
 		*/
 		private $sanitizedControllerName = '';
+
+		/**
+		* @var string Holds the sanitized class.
+		*/
+		private $sanitizedControllerClass = '';
 
 		public function __construct(string $string) {
 			$base = ucwords(preg_replace("/\W+/", ' ', strtolower($string)));
@@ -17,15 +22,21 @@ namespace Core {
 			$parts = array_map("ucfirst", $parts);
 			$base = implode('', $parts);
 
-			$this->sanitizedControllerName = "\\".$base . 'Controller';
+			$this->sanitizedControllerName = $base;
+
+			$this->sanitizedControllerClass = "\\".$base . 'Controller';
 		}
 
 		public function __toString() {
 			return $this->sanitizedControllerName;
 		}
 
-		public function getSanitiedControllerName() : string {
+		public function getSanitizedControllerName() : string {
 			return $this->sanitizedControllerName;
+		}
+
+		public function getSanitizedControllerClass() : string {
+			return $this->sanitizedControllerClass;
 		}
 	}
 }