+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog #
 v5.0.0
 - **This release is backwards incompatible.**  
+- Renamed the Document class to Assets  
+- Completely restructered the support for children controllers  
+- Controllers must now set their own view files.  
 - New file structure to enforce configuration files outside of the application directory.  
 - Removed Core\Configuration from the registry object, moved to Application::getConfiguration() and Theme::getConfiguration().  
 - Switched to .jsonc configuration file extension, to support comments in configuration files.